requirements.txt
absl-py @ file:///tmp/build/80754af9/absl-py_1607439979954/work
aiohttp @ file:///tmp/build/80754af9/aiohttp_1607109610110/work
albumentations==0.5.2
apex==0.1
appdirs==1.4.4
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
blinker==1.4
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1607706694405/work
certifi==2020.12.5
cffi @ file:///tmp/build/80754af9/cffi_1606255081583/work
chardet @ file:///tmp/build/80754af9/chardet_1605303185383/work
cityscapesScripts==2.1.7
click==7.1.2
cloudpickle==1.6.0
coloredlogs==15.0
cryptography @ file:///tmp/build/80754af9/cryptography_1607635341180/work
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
detectron2==0.3+cu102
dotmap==1.3.23
future==0.18.2
fvcore @ file:///home/conda/feedstock_root/build_artifacts/fvcore_1612446996969/work
google-auth @ file:///tmp/build/80754af9/google-auth_1604100403705/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1603929124518/work
grpcio @ file:///tmp/build/80754af9/grpcio_1597424474635/work
higher==0.2.1
humanfriendly==9.1
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
imageio==2.9.0
imgaug==0.4.0
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
iopath==0.1.3
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1604014535162/work
Markdown @ file:///tmp/build/80754af9/markdown_1605111056890/work
matplotlib @ file:///tmp/build/80754af9/matplotlib-base_1603378225747/work
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict @ file:///tmp/build/80754af9/multidict_1600456399709/work
networkx==2.5
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603570489231/work
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.46
opencv-python-headless==4.5.1.48
pandas @ file:///tmp/build/80754af9/pandas_1602088120436/work
panopticapi @ git+https://github.com/cocodataset/panopticapi.git@118ffaba478b5d45925e88d504625b4123f3160a
Pillow @ file:///tmp/build/80754af9/pillow_1603822255246/work
portalocker==2.0.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydot==1.4.1
pyjack==0.3.2
PyJWT==1.7.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1606517880428/work
pyparsing==2.4.7
pyquaternion==0.9.9
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.1
pytorch3d==0.3.0
pytz @ file:///tmp/build/80754af9/pytz_1606604771399/work
PyWavelets==1.1.1
PyYAML==5.3.1
requests @ file:///tmp/build/80754af9/requests_1606691187061/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1596998415516/work
scikit-image==0.18.1
scipy==1.4.1
seaborn @ file:///tmp/build/80754af9/seaborn_1600553570093/work
Shapely==1.7.1
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1605205327372/work
tabulate==0.8.7
tensorboard @ file:///home/builder/ktietz/conda/conda-bld/tensorboard_1604313476433/work/tmp_pip_dir
tensorboard-plugin-wit==1.6.0
tensorflow-gpu==2.2.0
termcolor==1.1.0
tifffile==2021.1.8
torch==1.7.0
torchaudio==0.7.0a0+ac17b64
torchvision==0.8.1
tornado @ file:///tmp/build/80754af9/tornado_1606942300299/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1610419111160/work
typing==3.7.4.3
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
Werkzeug==1.0.1
yacs==0.1.8
yarl @ file:///tmp/build/80754af9/yarl_1598045023180/work
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
